chore: refactor automaton, move dfa to module
